# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionProposalListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/ActionProposalResponse"
            },
            "title": "Proposals",
            "type": "array"
          }
        },
        "required": [
          "proposals",
          "count"
        ],
        "title": "ActionProposalListResponse",
        "type": "object"
      },
      "ActionProposalResponse": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalPreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "approved_execution_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionContextView"
              },
              {
                "type": "null"
              }
            ]
          },
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/ProposalAuditEvent"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ProposalEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "execution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionExecutionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution_context": {
            "$ref": "#/components/schemas/ExecutionContextView"
          },
          "execution_context_changed": {
            "title": "Execution Context Changed",
            "type": "boolean"
          },
          "execution_context_hash": {
            "title": "Execution Context Hash",
            "type": "string"
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "origin_fingerprint": {
            "title": "Origin Fingerprint",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/TypedActionPayload"
          },
          "payload_hash": {
            "title": "Payload Hash",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "simulation_only": {
            "title": "Simulation Only",
            "type": "boolean"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "origin_fingerprint",
          "action_type",
          "rationale",
          "source_refs",
          "evidence",
          "payload",
          "payload_hash",
          "version",
          "risk_level",
          "confidence",
          "status",
          "expires_at",
          "approval",
          "execution",
          "rejection_reason",
          "created_at",
          "updated_at",
          "audit_events",
          "execution_mode",
          "simulation_only",
          "execution_context",
          "execution_context_hash",
          "approved_execution_context",
          "execution_context_changed"
        ],
        "title": "ActionProposalResponse",
        "type": "object"
      },
      "ActionType": {
        "description": "Closed set of proposable actions. High-risk actions (send email,\ndelete event, purchase) are prohibited in the MVP and intentionally\nabsent \u2014 they cannot be proposed, approved, or executed.",
        "enum": [
          "create_task",
          "create_gmail_draft",
          "create_calendar_event"
        ],
        "title": "ActionType",
        "type": "string"
      },
      "ApprovalPreview": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "approved_at": {
            "format": "date-time",
            "title": "Approved At",
            "type": "string"
          },
          "binding_hash": {
            "title": "Binding Hash",
            "type": "string"
          },
          "execution_context": {
            "$ref": "#/components/schemas/ExecutionContextView"
          },
          "payload": {
            "$ref": "#/components/schemas/TypedActionPayload"
          },
          "payload_hash": {
            "title": "Payload Hash",
            "type": "string"
          },
          "proposal_version": {
            "title": "Proposal Version",
            "type": "integer"
          }
        },
        "required": [
          "action_type",
          "proposal_version",
          "payload",
          "payload_hash",
          "binding_hash",
          "approved_at",
          "execution_context"
        ],
        "title": "ApprovalPreview",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProposalApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "displayed_execution_context_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Displayed Execution Context Hash",
            "type": "string"
          },
          "displayed_payload_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Displayed Payload Hash",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version",
          "action_type",
          "displayed_payload_hash",
          "displayed_execution_context_hash"
        ],
        "title": "ProposalApprovalRequest",
        "type": "object"
      },
...
      "ProposalEditRequest": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskCreatePayload"
              },
              {
                "$ref": "#/components/schemas/GmailDraftCreatePayload"
              },
              {
                "$ref": "#/components/schemas/CalendarEventCreatePayload"
              }
            ],
            "title": "Payload"
          }
        },
        "required": [
          "expected_version",
          "action_type",
          "payload"
        ],
        "title": "ProposalEditRequest",
        "type": "object"
      },
...
      "ProposalRejectionRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "ProposalRejectionRequest",
        "type": "object"
      },
      "ProposalStatus": {
        "enum": [
          "proposed",
          "edited",
          "approved",
          "rejected",
          "executing",
          "executed",
          "failed",
          "expired"
        ],
        "title": "ProposalStatus",
        "type": "string"
      },
...
      "RiskLevel": {
        "enum": [
          "low",
          "medium"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LifeFlow AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/action-proposals": {
      "get": {
        "operationId": "list_action_proposals_action_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ProposalStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposalListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Action Proposals"
      }
    },
    "/action-proposals/{proposal_id}": {
      "get": {
        "operationId": "get_action_proposal_action_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Action Proposal"
      },
      "patch": {
        "operationId": "edit_action_proposal_action_proposals__proposal_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Action Proposal"
      }
    },
    "/action-proposals/{proposal_id}/approve": {
      "post": {
        "operationId": "approve_action_proposal_action_proposals__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Action Proposal"
      }
    },
    "/action-proposals/{proposal_id}/execute": {
      "post": {
        "operationId": "execute_action_proposal_action_proposals__proposal_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Action Proposal"
      }
    },
    "/action-proposals/{proposal_id}/reject": {
      "post": {
        "operationId": "reject_action_proposal_action_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRejectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Action Proposal"
      }
    },
...
  }
}